Changed links from an object to an array (#25)
Links is no longer an object with links expressed either as string or
as arrays of strings. Instead, links is an array, with each link
being represented by a single object. This object is a tuple of type
and target (an enumerated value and a UUID string, respectively).

Updated all mentions of links in the repository, including use cases
and their graphical representations.