Rewrote validation.py to perform schema validation (#49)
Instead of simply verifying JSON syntax by attempting to load the example files, validate.py now searches for schemas and for each schema applies it to any examples of the corresponding type.

A minor drawback of this is that since currently the set of schemas is incomplete, not all examples will be thus verified. This is only temporary, however.