Made DocumentationCreated rely on SourceChangeSubmitted
Following a workshop discussion regarding the ability of
DocumentationCreated to identify an issue (e.g. a bug report),
the proposal for DocumentationCreated has been updated.

Similarly to ConfigurationApplied (#79) it is longer
recommended only for cases where you do NOT store documentation
as source. Instead, it references SourceChangeSubmitted
using DOCUMENTATION_SOURCE.

To catch situations in general and for documentation in
particular, where these sources do not reside in e.g. Git
or SVN but in some other random, custom storage solution,
a data.universalResourceIdentifier property is added to
SourceChangeSubmitted and SourceChangeCreated as a catch-all
property.

The reason for spelling it out (rather than uri) is to
preserve likeness to all other ...Identifier properties,
and uriIdentifier would just be weird. Simply uri is used
in all other contexts, so it's a bit inconsistent, but
arguably the best option.