Made schemas version specific (#81)
As per issue #76, made each schema check the value of meta.version.
Consequently, each schema is now version specific. The directory
structure of schemas has been changed accordingly, and the
validation script now looks for a schema not just based on the
event type, but also based on the event version.