Updated schema validation to include zip archives
In order to ensure that reference data sets comply with scehma,
the validation script now extracts any zip archives and includes
any .json files within them.

As this can be very computationally demanding, additional arguments
have been added to the validation script to allow validation of a
(randomized) subset of all events.

The Travis CI script is unchanged, meaning that it will not check
archive contents.