Added license notice to all .md and .py files. (#133)
As per issue #130.

Added license notices to all .md and .py files and added instructions on how to keep this up to date in How to Contribute documentation. Notice was not added to .json and .gliffy files (which are also JSON files) as comments are not supported. Arguably that is less of a concern: the important thing is that documentation and code contains the license notice.