Added issue and PR templates
Created templates for issues and pull requests, as per issue #159.
The intention is for these templates to be minimalistic and easy
to use, yet guide contributors towards including sufficient
information so as to enable effective discussions and reviews.

The templates are intended to reflect the contribution guidelines,
while aligning to generally accepted good practice.