Stepped EiffelTestExecutionRecipeCollectionCreatedEvent to 2.0.0 (#155)
Stepped EiffelTestExecutionRecipeCollectionCreatedEvent to 2.0.0, as per issue #149.

In version 2.0.0 of the event, the syntax of data.batches.recipes.constraints changed to comply with protocol design guidelines. It is no longer an uncontrolled object, but a list of key-value pair, identical to data.customData.

Updated schema for EiffelTestExecutionRecipeCollectionCreatedEvent 2.0.0 and fixed incorrect UUID in EiffelTestExecutionRecipeCollectionCreatedEvent examples.