Moved documentation of link types to event type documentation files.
As described in issue #136, because of the per-event versioning scheme
it is untenable to keep link type documentation central and separate from
event types. When link types are changed, this constitutes an event
version change. Therefore, the link types of each event type must be
documented each respective event type documentation.

the-links-object.md has been preserved, but only contains general
information. The repository has been searched for occurrences of
the-links-object.md and has been updated accordingly.