Decentralized meta and links object documentation (Issue #136)
As described in issue #136, because of the per-event versioning scheme
it is untenable to keep link type documentation central and separate from
event types. When link types are changed, this constitutes an event
version change. Therefore, the link types of each event type must be
documented each respective event type documentation.

the-links-object.md has been preserved, but only contains general
information. The repository has been searched for occurrences of
the-links-object.md and has been updated accordingly.

As described by Issue #136, the per-event versioning scheme
requires that even common syntax be described on a per-event basis,
lest the documentation and/or versioning break down.

This commit moves the meta object documentation into the
documentation of each respective event.