Introduced purls instead of GAVs
Replaced GAVs for meta.source.serializer in all events,
and for data.identity (instead of data.gav) and data.implements
and data.dependsOn in EiffelArtifactCreatedEvent.

Updated schema and examples accordingly. In doing so, realized
that usage-examples/delivery-interface has not been updated
since EiffelIssueDefinedEvent was introduced, and so fixed this.

Gliffy is rather old and poorly maintained. Converted the
diagram source file from .gliffy to a draw.io .xml file.