All .gliffy files replaced by .svg (#199)
All .gliffy and .png files depicting event graphs replaced by .svg. Relevant .md files updated accordingly.

The test-execution-sequence diagram is preserved: this should eventually be replaced with a reference to Eiffel Sepia, but that requires rephrasing the example documentation and will have to wait.

Also fixed a number of typos and mistakes in the graphs.