Introduced purls instead of GAVs (#195)
Replaced GAVs for meta.source.serializer in all events, and for data.identity (instead of data.gav) and data.implements and data.dependsOn in EiffelArtifactCreatedEvent.

Updated schema and examples accordingly. As part of that change, replaced .gliffy files (dependent on a specific tool) with .svg files (tool agnostic) and fixed some existing issues with the images.

Replaced references to GAVs in compositions-and-validity-checking.md and confidence-level-joining.md. Also a typo was fixed in compositions-and-validity-checking.md.