Add {mediaType,tags} to data.{liveLogs,persistentLogs}
These new members for log location objects allows event producers to
describe the log locations in a more structured manner by including
tags and the media type of the log payload. This can e.g. be used to
distinguish between different representations of the same log (say,
HTML and plain text) or identify logs from a particular tool.

The schemas don't contain a regular expression for the media type member
and accept any string. This could be seen as a drawback, but media types
are pretty complicated beasts so instead of risking a regexp bug that
blocks legitimate media types we're lax about these strings, similar to
how ArtC's data.identity doesn't try to include a regular expression for
matching all package URLs.