Add domain id for links (#269)
To make cross-domain links unambiguous and make it easier and more
performant to locate the target events of such links we introduce a
domainId member to the link object. This is a backwards-compatible
change since the absence of the new member means that the target
event can be assumed to exist in the same domain as the source event.

Since we're now using the domain term in another place apart from the
description of meta.source.domainId in each event we centralize the
definition in a new glossary term.