tox: Upgrade Black dependency to fix Click incompatibility
Black depends on Click and seems to have use a non-public API. When
Click changed their API this broke Black, resulting in the following
error when running Black:

    ImportError: cannot import name '_unicodefun' from 'click' (...)

Upgrading Black to the most recent release addresses the problem.
We should be using Poetry to lock all transitive dependencies and
avoid this class of problems in the future.