Add test that verifies history tables entries (#352)
It's easy to forget to add an entry in the history table when you copy
an existing event definition file. We therefore add a Python test file
with a short test that verifies that each event definition has a
corresponding history table entry.

The new test exposed to definition files without history tables so we
had to correct them too.