Change definition of "introduced_in" field in the history tables
That field has previously only had a well-defined and obvious meaning
for event versions that were the latest version in a released edition.
To make the files more consistent and easier to maintain we change the
field to be the Git tag where the event version was first introduced,
or null if the version hasn't been released yet. This also makes the
field machine-readable, and we'll render it to a clickable link in the
Markdown to the documentation generator.