test_definitions.py: Load YAML files only once
Instead of loading the YAML files in @pytest.mark.parametrize
expression of each test case (i.e. we're parsing 169 YAML files
three times over), preload the files into a global variable that
we use in the test parametrization. This decreases the execution
time by about 65%.

Also, by passing the "ids" kwarg to @pytest.mark.parametrize we
can choose the name of each parametrized test so that's it's useful
description rather than just a generic string with an autoincremented
integer at the end. This allowed us to cut back on some assertion
failure messages since they no longer have to be so verbose.