Add _name and _version fields to all definition files (#392)
Up to now we've been relying on the directory structure and filenames
to determine the name and version of a definition file for an event
or some other type. There are advantages to this, but also the drawback
that it's clunky to work with and that the files always must remain in
this structure, i.e. you can't just transfer the contents of a file over
some channel or protocol and expect the receiver to be able to make
sense of it.

The new _name and _version key are introduced in the definition file
format and all existing files are updated to include them. We also add
a test that makes sure that the contents of those keys matches the
path to the file.