generate_manifest.py: Use versions module
When the versions module was introduced in commit e4762e9, the code in
generate_manifest.py for listing the latest version of each type went
unnoticed. With a slight change to versions.py to allow discovering
schema file and not just definitions files we were able to use that
module and delete some code.