Lock version of markdownlint-cli and fix violations
markdownlint-cli was using a floating version, and since this
linter was introduced additional rules have been added so now
the linting fails on the master branch. Lock the version to
the currently latest version and disable the rule that we
didn't comply with (since addressing it properly would be
rather intrusive).