Document the format of public keys and signatures (#414)
The meta.security.integrityProtection.{publicKey,signature} fields were 
underspecified, both in the documentation and the schema. We now
describe the expected format as well as state the formal requirements
for the Base64 strings.

The regular expression used for validating the strings is somewhat
sloppy and will allow some non-Base64 strings, but getting a 100%
correct expression isn't trivial and an expression that's too strict and 
rejects some rare corner case could cause real problems. There doesn't
appear to be a canonical regexp for Base64 strings, and relying on
StackOverflow posts with terribly complicated expressions doesn't feel
right.